add react router, timer and stopwatch routes, basic timer function
